plug-ins: Correct Legacy PSD layer style loading format
Per Wormnest, the 4 bytes allocated to
PSD Legacy Dropshadow properties are
meant to be used for fixed floating point
values rather than integers. This patch
corrects the PSD layer resource struct and
how the data is loaded, based on prior
work by Wormnest.

+++ plug-ins/file-psd/psd.h
@@ -593,10 +593,10 @@ typedef struct
 {
   guint32    size;
   guint32    ver;
-  guint16    blur;
-  guint32    intensity;
-  gint32     angle;
-  guint32    distance;
+  gfloat     blur;
+  gfloat     intensity;
+  gfloat     angle;
+  gfloat     distance;
   guint16    color[5];
   gchar      blendsig[4];
   guchar     effecton;
